+++ src/services/axios/postProposalReview.tsx/postProposalReview.test.tsx
@@ -7,11 +7,13 @@ import PostProposalReview, {
   postMockProposalReview
 } from './postProposalReview';
 import {
-  MockProposalReviewFrontend,
+  MockProposalScienceReviewExcludedFrontend,
+  MockProposalScienceReviewFrontend,
   MockProposalTechnicalReviewFrontend
 } from './mockProposalReviewFrontend';
 import {
-  MockProposalReviewBackend,
+  MockProposalScienceReviewBackend,
+  MockProposalScienceReviewExcludedBackend,
   MockProposalTechnicalReviewBackend
 } from './mockProposalReviewBackend';
 import { ProposalReviewBackend } from '@/utils/types/proposalReview';
@@ -35,11 +37,20 @@ describe('Helper Functions', () => {
 
   test('mappingReviewFrontendToBackend returns mapped review from frontend to backend format', () => {
     const reviewBackEnd: ProposalReviewBackend = mappingReviewFrontendToBackend(
-      MockProposalReviewFrontend,
+      MockProposalScienceReviewFrontend,
       cycleId,
       true
     );
-    expect(reviewBackEnd).to.deep.equal(MockProposalReviewBackend);
+    expect(reviewBackEnd).to.deep.equal(MockProposalScienceReviewBackend);
+  });
+
+  test('mappingReviewFrontendToBackend returns mapped review from frontend to backend format CHLOE', () => {
+    const reviewBackEnd: ProposalReviewBackend = mappingReviewFrontendToBackend(
+      MockProposalScienceReviewExcludedFrontend,
+      cycleId,
+      true
+    );
+    expect(reviewBackEnd).to.deep.equal(MockProposalScienceReviewExcludedBackend);
   });
 
   test('mappingReviewFrontendToBackend returns mapped technical review from frontend to backend format', () => {
@@ -54,14 +65,14 @@ describe('Helper Functions', () => {
   test('mappingReviewFrontendToBackend generates cycle id when not provided', () => {
     const reviewBackEnd: ProposalReviewBackend = mappingReviewFrontendToBackend(
       {
-        ...MockProposalReviewFrontend,
+        ...MockProposalScienceReviewFrontend,
         cycle: ''
       },
       cycleId,
       true
     );
     expect(reviewBackEnd).to.deep.equal({
-      ...MockProposalReviewBackend,
+      ...MockProposalScienceReviewBackend,
       cycle: 'SKAO_2027_1'
     });
   });
@@ -74,42 +85,42 @@ describe('PostProposalReview Service', () => {
 
   test('returns mock data id when USE_LOCAL_DATA is true', async () => {
     vi.spyOn(CONSTANTS, 'USE_LOCAL_DATA', 'get').mockReturnValue(true);
-    const result = await PostProposalReview(MockProposalReviewFrontend, cycleId);
+    const result = await PostProposalReview(MockProposalScienceReviewFrontend, cycleId);
     expect(result).toEqual('PROPOSAL-REVIEW-ID-001');
   });
 
   test('returns data id from API when USE_LOCAL_DATA is false', async () => {
     vi.spyOn(CONSTANTS, 'USE_LOCAL_DATA', 'get').mockReturnValue(false);
-    mockedAxios.post.mockResolvedValue({ data: MockProposalReviewBackend.review_id });
-    const result = (await PostProposalReview(MockProposalReviewFrontend, cycleId)) as string;
-    expect(result).to.deep.equal(MockProposalReviewBackend.review_id);
+    mockedAxios.post.mockResolvedValue({ data: MockProposalScienceReviewBackend.review_id });
+    const result = (await PostProposalReview(MockProposalScienceReviewFrontend, cycleId)) as string;
+    expect(result).to.deep.equal(MockProposalScienceReviewBackend.review_id);
   });
 
   test('returns error message on API failure', async () => {
     vi.spyOn(CONSTANTS, 'USE_LOCAL_DATA', 'get').mockReturnValue(false);
     mockedAxios.post.mockRejectedValue(new Error('Network Error'));
-    const result = await PostProposalReview(MockProposalReviewFrontend, cycleId);
+    const result = await PostProposalReview(MockProposalScienceReviewFrontend, cycleId);
     expect(result).toStrictEqual({ error: 'Network Error' });
   });
 
   test('returns error.API_UNKNOWN_ERROR when thrown error is not an instance of Error', async () => {
     vi.spyOn(CONSTANTS, 'USE_LOCAL_DATA', 'get').mockReturnValue(false);
     mockedAxios.post.mockRejectedValue({ unexpected: 'object' });
-    const result = await PostProposalReview(MockProposalReviewFrontend, cycleId);
+    const result = await PostProposalReview(MockProposalScienceReviewFrontend, cycleId);
     expect(result).toStrictEqual({ error: 'error.API_UNKNOWN_ERROR' });
   });
 
   test('returns error.API_UNKNOWN_ERROR when result undefined', async () => {
     vi.spyOn(CONSTANTS, 'USE_LOCAL_DATA', 'get').mockReturnValue(false);
     mockedAxios.post.mockResolvedValue(undefined);
-    const result = await PostProposalReview(MockProposalReviewFrontend, cycleId);
+    const result = await PostProposalReview(MockProposalScienceReviewFrontend, cycleId);
     expect(result).toStrictEqual({ error: 'error.API_UNKNOWN_ERROR' });
   });
 
   test('returns error.API_UNKNOWN_ERROR when result null', async () => {
     vi.spyOn(CONSTANTS, 'USE_LOCAL_DATA', 'get').mockReturnValue(false);
     mockedAxios.post.mockResolvedValue(null);
-    const result = await PostProposalReview(MockProposalReviewFrontend, cycleId);
+    const result = await PostProposalReview(MockProposalScienceReviewFrontend, cycleId);
     expect(result).toStrictEqual({ error: 'error.API_UNKNOWN_ERROR' });
   });
 });